feat: creating MigrationObjectPropertyWrapper class.
This class is meant to facilitate the task of obtaining the path taken to arrive at a property value. This is so that when a property is set, we don't just get it's value. We get this object which we can then use internally to obtain the path taken to use that value for assignment.