fix: `has_been_stored` gets called proactively and extensively to ensure we're not storing duplicate Migration Objects.
Since this involves calling the DB, if we get an object back (which means the object exists), we take advantage and populate some class params.

`id` => so that we have the Migration Object ID
`stored` => so that we know _not_ to try to store another of the same object.
`processed` => so that we know the object has already been processed. Absolutely essential for determining whether a Migration Object should be returned in something like an Unprocessed or Processed Data Chest list.