fix: ensure that wrapped MigrationObject is passed when properties are accessed.
When accessing a property on a wrapped MigrationObject, we were delegating the call to the MigrationObject itself. This caused a MigrationObjectPropertyWrapper to include/reference the unwrapped MigrationObject it came from. Instead of delegating the call, we need to directly access the property that is being queried, and if it exists, return a new instance of MigrationObjectPropertyWrapper with a reference to the wrapped MigrationObject.