feat: initial commit of FailedMigrationState class.
Created a FailedMigrationState class to help with recording of Errors/Exceptions that can occur during migration. In order to allow for customization of the FailedMigrationState class during runtime (which would need to be done within the `command` function), I needed to refactor classes so that we could get access to the MigrationRunContext class. This refactor, in my opinion, should absolutely be done on the main dev branch if we don't decide to go with this FailedMigrationSate class because it centralizes the way to obtain a MigrationRunKey. This is especially important, since MigrationState's could alter the MigrationRunKey when the `settle` function is called. Using `MigrationRunContext::get_run_key()` will ensure you are always getting the most current RunKey.