fix: adding namespace to de-duplicate migrations.
Migrations can have duplicate names (the value returned from `get_name`). There's too much overhead to run the `get_name` function for every migration to ensure that it's truly unique; so instead we're adding a new column to `migrations` which has to have the fully qualified name of the Migration class. Namespaces have to be unique, otherwise a PHP error is thrown. This way, we ensure that although the name can be duplicated, it is made unique by adding the namespace.