fix: need version option when obtaining latest migration status
If you just pass Migration Class and Name, you will get the latest status for the latest Migration Run. However, if you need the status for a previous Migration Run, it is necessary to give a version number. Otherwise, you'll only ever get the status for the latest Migration Run.