feat: using FailedMigrationState, also custom PHP shutdown function
The custom PHP shutdown function is essential so that the Migrator has an opportunity to record a FAILED status, and what the error was. WP's Error handler prevents us from touching a PHP Fatal error, so the flag `WP_DISABLE_FATAL_ERROR_HANDLER` is required when using this.

Surrounding almost the entire `start()` function in a try/catch will allow us to record the reason for any failure, record the failure message, and kill the script (if `should_stop() == true`).