feat: get `max_allowed_packet` size & ensure underlying data is smaller
AbstractRunAwareMigrationDataChest is where we have a direct link to the DB. This is where we check what the `max_allowed_packet` size is, and compare it against what the (perhaps estimated) byte size of the underlying data is. If the underlying data is larger than the max allowed packet size, then instead of saving that data in the DB, we save a hash of that data.